MAP api
This is NOT an updated version of login so don’t use it. We should
actually put it as a javascript event on main.php when the cardio
minutes pop up. This is just the code for the map in case i lose it or
delete so don’t use this, but look at the code…it works